Member API 연동
