+++ screens/Login/FindPasswordScreen.js
@@ -8,13 +8,10 @@ import {
 } from "react-native";
 import { theme } from "../../colors/color";
 import { useEffect, useState } from "react";
-import {
-  TopLayout,
-  JoinMemberStyle,
-  JoinMemberBtn,
-} from "../../component/JoinMemberComp";
+import { TopLayout, SignUpStyle, SignUpBtn } from "../../component/SignUpComp";
+import { changePassword } from "../../api/login/login";
 
-export default function JoinMemberPassword({ navigation, route }) {
+export default function FindPasswordScreen({ navigation, route }) {
   const [password, setPassword] = useState("");
   const [isValid, setIsValid] = useState(false);
   const [isCorrect, setIsCorrect] = useState(false);
@@ -53,25 +50,24 @@ export default function JoinMemberPassword({ navigation, route }) {
     const isMatched = text.length > 0 && password === text;
     setIsCorrect(isMatched);
 
-    if (isMatched) {
-      Keyboard.dismiss();
-    }
+    // if (isMatched) {
+    //   Keyboard.dismiss();
+    // }
   };
 
   return (
     <SafeAreaView style={theme.container}>
-      <View style={JoinMemberStyle.subContainer}>
+      <View style={SignUpStyle.subContainer}>
         {!isKeyboardVisible && (
           <TopLayout
-            title={"비밀번호를\n입력해주세요."}
-            subtext={"잊은 비밀번호는\n이메일로 다시 보내드릴게요."}
-            imageSource={require("../../assets/images/joinmembership/password_icon.png")}
+            title={"변경하실 새로운\n비밀번호를 입력해주세요."}
+            imageSource={require("../../assets/images/signup/password_icon.png")}
           />
         )}
         <ScrollView showsVerticalScrollIndicator={false}>
           <Text
             style={[
-              JoinMemberStyle.subText_black,
+              SignUpStyle.subText_black,
               { fontFamily: "Pretendard-SemiBold" },
             ]}
           >
@@ -84,13 +80,13 @@ export default function JoinMemberPassword({ navigation, route }) {
             keyboardType="default"
             autoCapitalize="none"
             style={{
-              ...JoinMemberStyle.textInputStyle,
+              ...SignUpStyle.textInputStyle,
               borderColor: isValid ? theme.main_blue : "#EDEDED",
             }}
           />
           <Text
             style={[
-              JoinMemberStyle.subText_black,
+              SignUpStyle.subText_black,
               { fontFamily: "Pretendard-SemiBold" },
             ]}
           >
@@ -104,20 +100,20 @@ export default function JoinMemberPassword({ navigation, route }) {
             autoCapitalize="none"
             secureTextEntry={true}
             style={{
-              ...JoinMemberStyle.textInputStyle,
+              ...SignUpStyle.textInputStyle,
               borderColor: isCorrect ? theme.main_blue : "#EDEDED",
             }}
           />
         </ScrollView>
       </View>
-      <JoinMemberBtn
+      <SignUpBtn
         nextCondition={isValid && isCorrect}
-        nextFunction={() =>
-          navigation.push("JoinMemberProfile", {
-            email: route.params.email,
-            password: password,
-          })
-        }
+        nextFunction={() => {
+          const isSuccess = changePassword(route.params.email, password);
+          if (isSuccess) {
+            navigation.replace("Success");
+          }
+        }}
         backText={"뒤로"}
       />
     </SafeAreaView>
